\name{zoo}
\alias{zoo}
\docType{data}
\title{
Zoo
}
\description{
A simple database containing 17 Boolean-valued attributes.
}
\usage{data(krvskp)}
\format{
  A data frame with 101 observations on 17 variables.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Zoo
}
\references{
Forsyth's PC/BEAGLE User's Guide.
}
\examples{
data(krvskp)
## maybe str(krvskp) ; plot(krvskp) ...
}
\keyword{datasets}
